+++ src/core/components.ts
@@ -17,6 +17,7 @@ import * as linkModule from '../components/link/linkDirective';
 import * as navBarModule from '../components/navbar/navbarDirective';
 import * as overlayModule from '../components/overlay/overlayDirective';
 import * as personacardModule from '../components/personacard/personacardDirective';
+import * as personaModule from '../components/persona/personaDirective';
 import * as progressIndicatorModule from '../components/progressindicator/progressIndicatorDirective';
 import * as searchboxModule from '../components/searchbox/searchboxDirective';
 import * as spinnerModule from '../components/spinner/spinnerDirective';
@@ -50,6 +51,7 @@ export var module: ng.IModule = ng.module('officeuifabric.components', [
   navBarModule.module.name,
   overlayModule.module.name,
   personacardModule.module.name,
+  personaModule.module.name,
   progressIndicatorModule.module.name,
   searchboxModule.module.name,
   spinnerModule.module.name,